requirements.txt
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=0.24.0
matplotlib>=3.4.0
seaborn>=0.11.0
pyyaml>=5.4.0
nfl_data_py==0.3.3
requests>=2.26.0
python-dotenv>=0.19.0
jupyter>=1.0.0
pyarrow>=7.0.0  # for parquet support
pytest>=7.0.0
black>=22.0.0
isort>=5.10.0
mypy>=0.950
pylint>=2.14.0 